Make two new gates compile on Linux — swift-corelibs createFile is not discardable and an offset literal needed its Int64
The pup gate builds tests with Swift 6.1.2 on Linux, where FileManager.createFile has no @discardableResult and the echo gate's offset expression did not infer Int64 inside the enum payload. Both are test-only; macOS never complained.

+++ Host/Tests/HostWireTests/HostileTimestampGateTests.swift
@@ -67,7 +67,7 @@ final class HostileTimestampGateTests: XCTestCase {
             try echo(seq: 0, t1: 999_999, t2: 40_000, t3: 40_500, at: 2_000),
             [.beaconEchoAccepted(
                 beaconSeq: 0,
-                offsetMicroseconds: (40_000 + 40_500 - 2_000) / 2,
+                offsetMicroseconds: Int64(40_000 + 40_500 - 2_000) / 2,
                 rttMicroseconds: 1_500)])
         XCTAssertEqual(host.session.srttMicroseconds, 1_500)
     }
